+++ src/registers/control.rs
@@ -311,22 +311,17 @@ mod x86_64 {
         ///
         /// The value is returned without a canonicality check. Use
         /// [`RawVirtAddr::try_into_48`] or [`RawVirtAddr::try_into_57`] to convert it into a
-        /// checked address type, depending on the paging mode of the kernel.
+        /// checked address type, depending on the paging mode of the kernel, or
+        /// [`RawVirtAddr::as_u64`] to get the raw value.
         #[inline]
         pub fn read() -> RawVirtAddr {
-            RawVirtAddr::new(Self::read_raw())
-        }
-
-        /// Read the current page fault linear address from the CR2 register as a raw `u64`.
-        #[inline]
-        pub fn read_raw() -> u64 {
             let value: u64;
 
             unsafe {
                 asm!("mov {}, cr2", out(reg) value, options(nomem, nostack, preserves_flags));
             }
 
-            value
+            RawVirtAddr::new(value)
         }
     }